feat: add support for writing ASCII tabs to a text file
Created a new module 'exporter.py' and added the argument --export.
It takes an optional output filename, and defaults to tabs.txt.
The exporter adds a Big Endian timestamp (YYYY.mm.dd) at the start
of the file.

Multiple tab generation will be implemented in the future. To prepare
for this change, the exporter takes tabs as a list.

Thus, assembled tabs are now added to a list 'generated_tabs' that
can be printed and exported.