feat(flask): add support for customizing chord generation
Added a form for customizing how many chord sets are generated, as well
as how many chords are included in each set.

routes.py now detects if the POST method is being used and captures the
values submitted in the form, then passes them to generate_chords().

This change achieves feature parity with the CLI version of the program.